requirements.txt
babel==2.17.0
blinker==1.9.0
click==8.1.8
docxcompose==1.4.0
docxtpl==0.19.1
Flask==3.1.0
itsdangerous==2.2.0
Jinja2==3.1.5
lxml==5.3.0
MarkupSafe==3.0.2
python-docx==1.1.2
setuptools==75.8.0
six==1.17.0
typing_extensions==4.12.2
Werkzeug==3.1.3
